move IEventFilter to ReaderIndex namespace
