+++ src/EventStore.ClientAPI.Embedded/FilteredEmbeddedSubscription.cs
@@ -6,6 +6,7 @@ using EventStore.Core.Bus;
 using EventStore.Core.Data;
 using EventStore.Core.Messages;
 using EventStore.Core.Messaging;
+using EventStore.Core.Services.Storage.ReaderIndex;
 using EventStore.Core.Util;
 
 namespace EventStore.ClientAPI.Embedded {
