+++ src/EventStore.Core.Tests/Services/Storage/AllReader/when_reading_all_with_filtering.cs
@@ -2,6 +2,7 @@
 using NUnit.Framework;
 using EventStore.Core.Data;
 using EventStore.Core.Messages;
+using EventStore.Core.Services.Storage.ReaderIndex;
 using EventStore.Core.Util;
 using static EventStore.Core.Messages.TcpClientMessageDto.Filter;
