+++ src/EventStore.Core.Tests/Services/Transport/Grpc/AnyStreamRevisionTests.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using EventStore.Client;
 using EventStore.Core.Services.Transport.Grpc;
 using NUnit.Framework;
