+++ src/EventStore.Core/Services/Storage/ReaderIndex/IEventFilter.cs
@@ -1,6 +1,6 @@
 using EventStore.Core.Data;
 
-namespace EventStore.Core.Util {
+namespace EventStore.Core.Services.Storage.ReaderIndex {
 	public interface IEventFilter {
 		bool IsEventAllowed(EventRecord eventRecord);
 	}
