+++ src/EventStore.Core/Services/Transport/Tcp/ClientTcpDispatcher.cs
@@ -6,6 +6,7 @@ using EventStore.Core.Data;
 using EventStore.Core.Helpers;
 using EventStore.Core.Messages;
 using EventStore.Core.Messaging;
+using EventStore.Core.Services.Storage.ReaderIndex;
 using EventStore.Core.Util;
 
 namespace EventStore.Core.Services.Transport.Tcp {
